changed some size_t to int, added Roman numerals

+++ euler/functions/factorial.cc
@@ -1,7 +1,7 @@
 #include "../euler.h"
 
 // Returns factorial of number.
-double euler::factorial(double num)
+double euler::factorial(int num)
 {
     if (num > 1)
         return num * factorial(num - 1);
